fix(manager): merge useField and useSubscription
